+++ src/main/kotlin/com/algorist/zMyBatis/startup/MyBatisActionInterceptorActivity.kt
@@ -5,7 +5,6 @@ import com.intellij.database.console.JdbcConsole
 import com.intellij.database.util.DasUtil
 import com.intellij.database.util.ObjectPath
 import com.intellij.database.util.SearchPath
-import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.fileTypes.FileTypeManager
@@ -24,12 +23,6 @@ class MyBatisActionInterceptorActivity : ProjectActivity {
     override suspend fun execute(project: Project) {
         LOG.info("zMyBatis: startup activity running")
 
-        // Set the icon for zMyBatis.ExplainGroup. The icon= attribute in plugin.xml only
-        // accepts resource paths, not class references, so we set it programmatically here.
-        ActionManager.getInstance()
-            .getAction("zMyBatis.ExplainGroup")
-            ?.templatePresentation
-            ?.icon = icons.DatabaseIcons.ConsoleShowPlan
 
         // Register the project-close listener BEFORE scheduling restoreSessionsIntoCache.
         // If the listener were registered inside the invokeLater lambda there would be a